+++ src/api/users.rs
@@ -1,9 +1,8 @@
 use rocket::serde::json::Json;
 
-use crate::auth::HaasUser;
 use crate::models::user::User;
 
 #[get("/users/me")]
-pub fn me(user: HaasUser) -> Json<User> {
-    Json(user.0)
+pub fn me(user: User) -> Json<User> {
+    Json(user)
 }
